Initial commit
This commit actually contains days of work, leading to something like
a 0.0.4 version – an app that works, but is still incomplete and in
progress.

The reason for this initial commit being a big dump of code is that,
unusually for me, all the work was done without caring much about the
commit history so far – I was basically using commits like savepoints,
instead of taking the time to explain each change, as I should have.

Therefore, I've decided to just smash everything together and maybe,
now, start creating correct commits. Hopefully.