Make the consumable/condition illustration a btn
Using CSS generated content for the illustration of each condition
(sleepless, cold…) or consumable (torches, arrows…) was nice, since
these images don't add any semantic value. But, because I _also_ wanted
to apply an image filter on them in certain circumstances, they couldn't
be generated as _background_ images. Which, in turn, led to other issues,
because images as generated content keep their initial size as far as
layout computation is concerned, even when scaled. So they had to be
_absolutely_ placed, and this made aligning the different conditions
and consumables difficult.

On top of that, we need to provide the players with a way to toggle
conditions, and decrease or increase their consumables. So I've decided
to kill 2 birds with 1 stone:

- Add a button for the required action.
- Put the image on this button instead of in a generated content.

This does mean that, even though semantically irrelevant, the image
will be in the DOM, but it's an acceptable trade-off. And it makes it
possible to simplify the CSS layout so that everything is well aligned.