Added `characters/conditions#toggle`
This new, non-RESTful action simply toggles the requested “condition”
attribute (`Character#sleepless`, `Character#thirsty`,
`Character#hungry` or `Character#cold`).

Note that I chose _not_ to go strictly RESTful with a
`ConditionTogglesController#create`, because it's such a small action
that it felt overkill. But this could change, especially if we were to
introduce notifications or somehow treat this action as a more relevant
_event_ – as increasing or decreasing willpower if, for example.